refactor: big Dropdown refactor
- Dropdown elements are now created by extended classes from
AppElementCreator with AppElementUI class methods and fields

- grouping of UI classes in corresponding directories

- separate form-data classes into separate files
