+++ firebase-debug.log
@@ -731,3 +731,21 @@ Issues? Report them at https://github.com/firebase/firebase-tools/issues and att
 [info] i  functions: Finished "app" in ~1s {"metadata":{"emulator":{"name":"functions"},"function":{"name":"app"},"message":"Finished \"app\" in ~1s"}}
 [debug] [2021-03-27T14:54:37.777Z] [worker-app-ffac0cdb-8c0d-401d-b6d9-433a5321da9c]: IDLE {"metadata":{"emulator":{"name":"functions"},"message":"[worker-app-ffac0cdb-8c0d-401d-b6d9-433a5321da9c]: IDLE"}}
 [debug] [2021-03-27T14:54:37.777Z] [work-queue] {"queueLength":0,"workRunningCount":0}
+[debug] [2021-03-27T14:55:31.835Z] [work-queue] {"queueLength":1,"workRunningCount":0}
+[debug] [2021-03-27T14:55:31.835Z] [work-queue] {"queueLength":0,"workRunningCount":1}
+[debug] [2021-03-27T14:55:31.835Z] Accepted request GET /save-and-save/us-central1/app/api/reports --> app
+[debug] [2021-03-27T14:55:31.836Z] [worker-pool] submitWork(triggerId=app) {"metadata":{"emulator":{"name":"functions"},"message":"[worker-pool] submitWork(triggerId=app)"}}
+[debug] [2021-03-27T14:55:31.836Z] [worker-app-ffac0cdb-8c0d-401d-b6d9-433a5321da9c]: Assigning socketPath: /var/folders/rd/b8y0hbfn6px0034frq1clv6m0000gn/T/fire_emu_13965.sock {"metadata":{"emulator":{"name":"functions"},"message":"[worker-app-ffac0cdb-8c0d-401d-b6d9-433a5321da9c]: Assigning socketPath: /var/folders/rd/b8y0hbfn6px0034frq1clv6m0000gn/T/fire_emu_13965.sock"}}
+[debug] [2021-03-27T14:55:31.836Z] [worker-app-ffac0cdb-8c0d-401d-b6d9-433a5321da9c]: BUSY {"metadata":{"emulator":{"name":"functions"},"message":"[worker-app-ffac0cdb-8c0d-401d-b6d9-433a5321da9c]: BUSY"}}
+[debug] [2021-03-27T14:55:31.837Z] [runtime-status] [13965] Trigger "app" has been found, beginning invocation! {"metadata":{"emulator":{"name":"functions"},"function":{"name":"app"},"message":"[runtime-status] [13965] Trigger \"app\" has been found, beginning invocation!"}}
+[info] i  functions: Beginning execution of "app" {"metadata":{"emulator":{"name":"functions"},"function":{"name":"app"},"message":"Beginning execution of \"app\""}}
+[debug] [2021-03-27T14:55:31.837Z] [runtime-status] [13965] triggerDefinition {"httpsTrigger":{},"name":"app","entryPoint":"app"} {"metadata":{"emulator":{"name":"functions"},"function":{"name":"app"},"message":"[runtime-status] [13965] triggerDefinition {\"httpsTrigger\":{},\"name\":\"app\",\"entryPoint\":\"app\"}"}}
+[debug] [2021-03-27T14:55:31.837Z] [runtime-status] [13965] Running app in mode HTTPS {"metadata":{"emulator":{"name":"functions"},"function":{"name":"app"},"message":"[runtime-status] [13965] Running app in mode HTTPS"}}
+[debug] [2021-03-27T14:55:31.838Z] [runtime-status] [13965] Attempting to listen to socketPath: /var/folders/rd/b8y0hbfn6px0034frq1clv6m0000gn/T/fire_emu_13965.sock {"metadata":{"emulator":{"name":"functions"},"function":{"name":"app"},"message":"[runtime-status] [13965] Attempting to listen to socketPath: /var/folders/rd/b8y0hbfn6px0034frq1clv6m0000gn/T/fire_emu_13965.sock"}}
+[debug] [2021-03-27T14:55:31.838Z] [functions] Runtime ready! Sending request! {"metadata":{"emulator":{"name":"functions"},"message":"[functions] Runtime ready! Sending request!"}}
+[debug] [2021-03-27T14:55:31.838Z] [functions] Got req.url=/save-and-save/us-central1/app/api/reports, mapping to path=/api/reports {"metadata":{"emulator":{"name":"functions"},"message":"[functions] Got req.url=/save-and-save/us-central1/app/api/reports, mapping to path=/api/reports"}}
+[debug] [2021-03-27T14:55:31.840Z] [runtime-status] [13965] Ephemeral server handling GET request {"metadata":{"emulator":{"name":"functions"},"function":{"name":"app"},"message":"[runtime-status] [13965] Ephemeral server handling GET request"}}
+[debug] [2021-03-27T14:55:31.840Z] [runtime-status] [13965] Ephemeral server survived. {"metadata":{"emulator":{"name":"functions"},"function":{"name":"app"},"message":"[runtime-status] [13965] Ephemeral server survived."}}
+[info] i  functions: Finished "app" in ~1s {"metadata":{"emulator":{"name":"functions"},"function":{"name":"app"},"message":"Finished \"app\" in ~1s"}}
+[debug] [2021-03-27T14:55:31.949Z] [worker-app-ffac0cdb-8c0d-401d-b6d9-433a5321da9c]: IDLE {"metadata":{"emulator":{"name":"functions"},"message":"[worker-app-ffac0cdb-8c0d-401d-b6d9-433a5321da9c]: IDLE"}}
+[debug] [2021-03-27T14:55:31.949Z] [work-queue] {"queueLength":0,"workRunningCount":0}
