+++ js/classes/app-form-data/AmountOfInjures.ts
