+++ js/classes/app-form-data/InjuredFromAttack.ts
