+++ js/components/redirected.ts
@@ -1,10 +1,10 @@
 import { api } from "../api/api";
-import AppTable from "../classes/AppTable";
+import AppTable from "../classes/app-table/AppTable";
 import { TableElements, TableElementProps } from "../interfaces/appTable"; 
-import FormData from "../classes/form-data/FormData";
-import InjuredFromAttack  from "../classes/form-data/InjuredFromAttack";
-import InjureType from "../classes/form-data/InjureType";
-import AmountOfInjures from "../classes/form-data/AmountOfInjures";
+import FormData from "../classes/app-form-data/FormData";
+import InjuredFromAttack  from "../classes/app-form-data/InjuredFromAttack";
+import InjureType from "../classes/app-form-data/InjureType";
+import AmountOfInjures from "../classes/app-form-data/AmountOfInjures";
 import downloadPdf from "../utils/createPdf";
